onBoard and auth screens

+++ app/src/main/java/com/elitec/satexplorer/feature/tracking/domain/entity/TleData.kt
@@ -3,7 +3,7 @@ package com.elitec.satexplorer.feature.tracking.domain.entity
 data class TleData(
     val line1: String,
     val line2: String,
-    val epoch: Double,
+    val epoch: Long,
     val meanMotion: Double,
     val eccentricity: Double,
     val inclination: Double,
